[BGD11] Timeline changes

+++ Goles/ViewControllers/TimeLine/TimeLineViewController.m
@@ -160,7 +160,7 @@
 }
 
 //------------------------------------------------------------------------------
-- (CGFloat)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath{
+- (float)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath{
 
     Shot *shot = self.arrayShots[indexPath.row];
